backend/requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.2.1
gunicorn==19.9.0
kombu==4.5.0
pillow==6.2.1
psycopg2==2.8.4
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
vine==1.3.0
django-cors-headers==3.7.0
django-redis==4.10.0
djangorestframework==3.9.4
mysqlclient==2.0.3
django-mysql==3.9
beautifulsoup4==4.9.3
bs4==0.0.1
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
django-rest-passwordreset==1.2.0